moved main window to its own class and fixed the scrollbar appearing when not needed
